refactor: split GameService turn-trace AI sections and map-view serialization (#3878)
Extract AI trace order aggregation into game_service_turn_trace_ai_sections.dart
and split map_view_serialization into read/write library parts to keep each
file under the repo file-size targets.
